+++ test/environmentSettings.ts
@@ -18,7 +18,6 @@ export class EnvironmentSettings {
     // OR
     // './lib/specs/**/anInitialSuite.e2e.js',
     // './lib/specs/**/apexLsp.e2e.js',
-    // './lib/specs/**/apexLspRestart.e2e.js',
     // './lib/specs/**/apexReplayDebugger.e2e.js',
     // './lib/specs/**/auraLsp.e2e.js',
     // './lib/specs/**/authentication.e2e.js',
